requirements.txt
ipython==7.5.0
PyYAML==5.1.1
PySide2==5.12.4
pyqtgraph==0.10.0
PyOpenGL==3.1.0
matplotlib==3.1.0
numpy==1.16.4
scipy==1.3.0
